+++ test/e2e/multikueue/baseline/suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package multikueue
+package baseline
 
 import (
 	"cmp"
